Revert "Improve some code, fix some bugs (#3317)"
This reverts commit 0a0014b5fc7c8e7cfdd928dbf7cab6810e14097a.

+++ extension/src/main/java/com/alibaba/fastjson2/support/AwtRederModule.java
@@ -72,7 +72,7 @@ public class AwtRederModule
         public Color apply(Map<Long, Object> values) {
             Integer rgb = (Integer) values.get(HASH_RGB);
             if (rgb != null) {
-                return new Color(rgb);
+                return new Color(rgb.intValue());
             }
 
             Integer r = (Integer) values.get(HASH_R);
